+++ src/ReceivingList/ExportSettingsModal/utils/isAllSelected.js
@@ -1,0 +1,3 @@
+import { EXPORT_FIELDS_PARAMS } from '../constants';
+
+export const isAllSelected = (param) => param === EXPORT_FIELDS_PARAMS.all;