+++ src/ReceivingList/ReceivingList.js
@@ -27,14 +27,16 @@ import {
   SingleSearchForm,
   PrevNextPagination,
   useFiltersToogle,
+  useItemToView,
   useLocationFilters,
   useLocationSorting,
-  useItemToView,
+  useModalToggle,
 } from '@folio/stripes-acq-components';
 
 import TitleDetailsContainer from '../TitleDetails';
+import { ExportSettingsModal } from './ExportSettingsModal';
+import { ReceivingListActionMenu } from './ReceivingListActionMenu';
 import ReceivingListFilter from './ReceivingListFilter';
-import { renderNewButton } from './renderNewButton';
 import {
   searchableIndexes,
 } from './ReceivingListSearchConfig';
@@ -96,8 +98,17 @@ const ReceivingList = ({
     changeSorting,
   ] = useLocationSorting(location, history, resetData, sortableFields);
   const { isFiltersOpened, toggleFilters } = useFiltersToogle('ui-receiving/filters');
+  const [isExportModalOpened, toggleExportModal] = useModalToggle();
 
-  const renderLastMenu = useCallback(() => renderNewButton(location.search), [location.search]);
+  const renderLastMenu = useCallback(({ onToggle }) => {
+    return (
+      <ReceivingListActionMenu
+        onToggle={onToggle}
+        titlesCount={titlesCount}
+        toggleExportModal={toggleExportModal}
+      />
+    );
+  }, [titlesCount, toggleExportModal]);
 
   const selectedTitle = useCallback(
     (e, { id }) => {
@@ -177,7 +188,7 @@ const ReceivingList = ({
           autosize
           title={resultsPaneTitle}
           count={titlesCount}
-          renderLastMenu={renderLastMenu}
+          renderActionMenu={renderLastMenu}
           toggleFiltersPane={toggleFilters}
           filters={filters}
           isFiltersOpened={isFiltersOpened}
@@ -218,6 +229,12 @@ const ReceivingList = ({
             </>
           ))}
         </ResultsPane>
+
+        {isExportModalOpened && (
+          <ExportSettingsModal
+            onCancel={toggleExportModal}
+          />
+        )}
 
         <Route
           path={`${match.path}/:id/view`}